Revamp event logic (#14)
Reduce number of events being simulated. The main changes is the removal of PollQueueEvent and compressing of QueueEvents.

Removal of EventUtil class as it only contained the convertPayloadsToQueueEvents which is now part of QueueEvent logic.
Updated node time comparison to use Double::compare.
Move setOccupied logic out of ProcessingDelayEvent to events that designate the start of processing which are ProcessedPayloadEvent and QueueEvent.
Moved original PollQueueEvent logic into both ProcessedPayloadEvent and QueueEvent.