Fix Round Change backlog bug (#16)
There was a bug when running IBFT on a topology where a round's RC message could arrive late to another validator after already moving to the next consensus instance. This causes an NPE. This has been fixed and it assumes the correctness of the backlog-dropping mechanism, which implies that one where the backlog pulls up a null for that current consensus instance implies it is outdated.